refactor: Move EmptyContent and SpinnerOrEmptyContent into folder
no code change
